Move 'static' pkg one level up.
