+++ static/analysis.go
