+++ static/app_test.go
@@ -3,7 +3,7 @@ package static_test
 import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/configkit/internal/entity"
-	. "github.com/dogmatiq/configkit/internal/static"
+	. "github.com/dogmatiq/configkit/static"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"golang.org/x/tools/go/packages"
@@ -28,7 +28,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app>",
 						Key:  "8a6baab1-ee64-402e-a081-e43f4bebc243",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/apps/single-app.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/apps/single-app.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -57,7 +57,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app-first>",
 						Key:  "b754902b-47c8-48fc-84d2-d920c9cbdaec",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/apps/multiple-apps-in-pkgs/first.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/apps/multiple-apps-in-pkgs/first.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -69,7 +69,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app-second>",
 						Key:  "bfaf2a16-23a0-495d-8098-051d77635822",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/apps/multiple-apps-in-pkgs/second.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/apps/multiple-apps-in-pkgs/second.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -98,7 +98,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app-first>",
 						Key:  "4fec74a1-6ed4-46f4-8417-01e0910be8f1",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/apps/multiple-apps-in-single-pkg/apps.AppFirst",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/apps/multiple-apps-in-single-pkg/apps.AppFirst",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -110,7 +110,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app-second>",
 						Key:  "6e97d403-3cb8-4a59-a7ec-74e8e219a7bc",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/apps/multiple-apps-in-single-pkg/apps.AppSecond",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/apps/multiple-apps-in-single-pkg/apps.AppSecond",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -139,7 +139,7 @@ var _ = Describe("func FromPackages() (application detection)", func() {
 						Name: "<app>",
 						Key:  "b754902b-47c8-48fc-84d2-d920c9cbdaec",
 					},
-					TypeNameValue: "*github.com/dogmatiq/configkit/internal/static/testdata/apps/pointer-receiver-app.App",
+					TypeNameValue: "*github.com/dogmatiq/configkit/static/testdata/apps/pointer-receiver-app.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
