+++ static/ident_test.go
@@ -3,7 +3,7 @@ package static_test
 import (
 	"github.com/dogmatiq/configkit"
 	"github.com/dogmatiq/configkit/internal/entity"
-	. "github.com/dogmatiq/configkit/internal/static"
+	. "github.com/dogmatiq/configkit/static"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"golang.org/x/tools/go/packages"
@@ -28,7 +28,7 @@ var _ = Describe("func FromPackages() (application identity)", func() {
 						Name: "<app>",
 						Key:  "04e12cf2-3c66-4414-9203-e045ddbe02c7",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/ident/const-value-ident.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/ident/const-value-ident.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -57,7 +57,7 @@ var _ = Describe("func FromPackages() (application identity)", func() {
 						Name: "<app>",
 						Key:  "9d0af85d-f506-4742-b676-ce87730bb1a0",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/ident/literal-value-ident.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/ident/literal-value-ident.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
@@ -86,7 +86,7 @@ var _ = Describe("func FromPackages() (application identity)", func() {
 						Name: "",
 						Key:  "",
 					},
-					TypeNameValue: "github.com/dogmatiq/configkit/internal/static/testdata/ident/variable-value-ident.App",
+					TypeNameValue: "github.com/dogmatiq/configkit/static/testdata/ident/variable-value-ident.App",
 					MessageNamesValue: configkit.EntityMessageNames{
 						Produced: nil,
 						Consumed: nil,
