+++ static/static.go
