+++ static/testdata/apps/multiple-apps-in-pkgs/first/app.go
