+++ static/testdata/apps/multiple-apps-in-pkgs/second/app.go
