+++ static/testdata/apps/multiple-apps-in-single-pkg/apps/first.go
