+++ static/testdata/apps/no-app/app.go
