+++ static/testdata/apps/pointer-receiver-app/app.go
