+++ static/testdata/apps/single-app/app.go
