+++ static/testdata/ident/const-value-ident/app.go
