+++ static/testdata/ident/literal-value-ident/app.go
