+++ static/testdata/ident/variable-value-ident/app.go
