+++ static/testdata/invalid/invalid-syntax/invalid.go
